backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.3
python-dotenv==1.0.1
Pillow==10.4.0
resend>=2.0.0
itsdangerous>=2.0.0
